Merge pull request #1802 from RosettaCommons/JWLabonte/sugars/glycomutagenesis
Shotgun Glycomutagenesis

This merge will introduce an application for shotgun glycomutagenesis, using the `GlycosyltransferaseMover`.

It depends on PRs #1786 and #1537.

The code is hacky, in my opinion, but that's why it's a pilot app, not a public app. I also need some of the underlying `EnzymeMover` code in master so I can complete and expand it for other PTMs for my student's research.

This merge also fixes a bug in which `-no_pro_close_ring_closure` was turned on by default when `-include_sugars` was on. (I think we do want `-no_pro_close_ring_closure` on ultimately, but only once we ensure that the other ring close method is being used as well.)

Because of the bug fix, all tests using the `-include_sugars` flag will have minor scoring differences. The unit test timeouts are a known major issue unrelated to this PR.

Old repository SHA1: 8a72a744464b85a068b5c11b237b9b8261761c95