Merge pull request #3845 from RosettaCommons/BYachnin/mhc_epitope-map-predictor
Added a "PreLoaded" MHCEpitopePredictor, analogous to "External" but held entirely in memory

MHCEpitopeEnergy currently allows for two Predictor classes that are used to obtain epitope scores from peptides: Matrix, which uses the Propred matrices to score peptides, and External, which uses an external SQL database (located on disk) to score peptides. This database is generated by the user based on the epitope prediction method and design space to be explored in a particular protein.

This PR provides functionality analogous to "External," but loads the database entirely into memory prior to scoring. In cases where the database is sufficiently small (i.e. will not use too much RAM), this will be faster than External, particularly in cases where multiple processes are accessing the same database file during packing. In addition, the ability to read "CSV" files for this purpose is enabled.

We also have introduced the database "unseen handler" (i.e. how to deal with peptides that are not in the database) "ignore," which will score unknown peptides with "0." This is suited for a database containing the IEDB experimentally-validated list of epitopes, which should introduce penalties for recognized peptides and no penalty for other peptides.

Integration test failure in mhc_epitope is from cosmetic changes to the tracer output.  mhc_epitope_nmer_preload is a new integration test, and so it fails.  recon_design_mpi is unstable in master, and continuous_sewing_hasher and discontinuous_sewing_hasher fail in master.

Old repository SHA1: 9a53bc0398564f83cfc18289706363d18c0672af

Co-authored-by: Chris Bailey-Kellogg <cbk@cs.dartmouth.edu>