Merge pull request #3876 from RosettaCommons/vmullig/fix_peptoid_clustering
Fix a bug with energy-based clustering and N-methylation

The alignment atoms were not being chosen properly in N-methylated peptides. Note that this was early code for me, and is very badly written -- it's much more delicate than it needs to be. I don't have time to rewrite it from scratch, but I'm adding a check that will at least produce an exception if the problem recurs.

Details:

I have a vector of atoms to be used in the RMSD, and in N-methyl amino acids, you may or may not want to use the carbon on the nitrogen in the RMSD. I had one place in my code where it was assumed that you would, and another where it was assumed that you wouldn't. The effect was that a reference structure's vector of atom coordinates had 5*N entries, the structure to be aligned had a vector with 5*N entries in which only the first 4*N entries were being populated, and the two vectors didn't line up correctly.

There are obviously much less fragile ways to write this that don't involve keeping track if which vector index corresponds to what, but like I said, this was early code for me and rewriting it entirely would take a lot of time.

Old repository SHA1: 79404f570f4d73df3cba84be7d1a4888281397ab