Merge pull request #3900 from RosettaCommons/roccomoretti/fix_ForceDisulfidesMover
Fixes for ForceDisulfidesMover

A previous PR (#3841) actually broke the ForceDisulfidesMover (due to how the regex patterns weren't nested.) Fix that. Also, add unit tests such that we don't do that in the future. 

Also, make a stab at the ability to remove disulfide bonds through the RosettaScripts interface. The current implementation is somewhat limited - it simply allows you to use the ForceDisulfidesMover to reset/remove all the existing disulfides prior to adding on the explicitly listed ones (can be used with a missing disulfides list to remove all of them).

I also added general functionality to allow people to unit test script validation for Movers (as the regex error is only caught by XSD validation, and not parse_my_tag). It works, but it's not as general as I'd like, due to the Mover special casing needed for adding the top-level element.

Old repository SHA1: e82622ca9b2aff0ab2f8e92cd55d21e49b18290a