Merge pull request #3909 from RosettaCommons/vmullig/fix_pose_copy_error
Fix a really stupid mistake that I introduced in packer cleanup with nondecomposible scoreterms. A cleanup function was receiving a copy of a pose instead of an instance of the pose to clean up. So it would nicely clean up the copy, then promptly delete it.

Old repository SHA1: 45d684cd82af6bf22c006738e29b46d019a71e81