Merge pull request #3925 from RosettaCommons/BYachnin/fix_buns_reporter_handling
The documentation and code comments for the BuriedUnsatisfiedHBonds filter loudly state that if no reporter behaviours are selected, `report_all_heavy_atom_unsats` will be set to true by default. If we actually run it this way, though, it returns 911. I've changed the default behaviour in parse_my_tag, and also in the compute function, as a fallback for PyRosetta/C++, to act as though `report_all_heavy_atom_unsats` is set to true.

The ppi_v3_suiteA tests will fail.  These tests were using the BUnS filter using the default settings, and getting the "911" result for the error case of now reporters being set to true.  This is now corrected, with a reasonable value being output as the BUnS result.  Other test failures appear to be timeouts or unstable.

Old repository SHA1: 39a6c320681c5347a88dc31316cab45b4c4d965f