Merge pull request #3938 from RosettaCommons/vmullig/no_thread_local
Replace an instance of thread_local with THREAD_LOCAL.

The thread_local keyword is not supported by all compilers. We use a custom THREAD_LOCAL macro which is defined as thread_local in those compilation environments that support the keyword, and as nothing in those that do not.

Old repository SHA1: 49333172bb4614ce8226cce4e573ed7b0b4a79d0