Merge pull request #3952 from RosettaCommons/roccomoretti/fix_stderr_redirect
Fix stderr redirect issues.

The order in which you put the "2>&1" stderr redirection matters. The shell parses output redirection from left to right, and the "2>&1" tells the shell to redirect stderr ("2") to the current location of stdout ("1"), not to make them match in perpetuity. So if you redirect output to a file, you need `> logfile 2>&1`, not `2>&1 > logfile`.

As such, fix a number of locations (primarily in testing) which were putting things the wrong way around. 

Old repository SHA1: 8b7ef155363009ec0a01d65cfb1e4e182b06f06e