Merge pull request #3933 from RosettaCommons/JWLabonte/sugars/OGT
Glycosylation: Adding EnzymaticMover for O-linked glycosylation

This merge primarily adds an O-linking glycosyltransferase to the database for use with the `GlycosyltransferaseMover`.

It also moves some testing apps to public, so that people can start using `EnzymaticMover`s more readily.

It also allows an optional string `-` to be used to tell the `EnzymaticMover` system to modify at the default atom for a `ResidueType`. This is helpful for cases with ambiguous site modifications, such as either Ser or Thr, which each have a different name for their hydroxyl oxygen in their `.params` files.

Finally, it fixes a bug in `GlycanTree`.

Documentation for the two added apps is available on the documentation wiki.

All unit tests pass. The two integration test changes are expected.

Old repository SHA1: 3301854f3b5524991b07e1cf8783857aa79d84b9

Co-authored-by: Jared Adolf-Bryfogle <jadolfbr@gmail.com>