Merge pull request #4036 from RosettaCommons/roccomoretti/integration_python3
Switch the test server to run integration.py with Python3.

Use Python3 on the test server (the same one running the benchmark.py script), such that Sergey only has to manage dependencies for one(ish) Python version.

While at this point we should be able to run integration.py with either Python2 or Python3, not all of the sub-scripts being run in the command files of the integration tests themselves can do so. To address this, add python2 and python3 substitution variables in the integration test scripts (to go along with the existing python one), for those tests which need a specific version. -- Please continue to use `%(python)s` if it the 2/3 choice doesn't matter - this will be whatever python with which the integration.py script was launched. The interpreter for python2 or python3 may be autodiscovered based on your path, and might not be the one you really want to use.

This also updates the integration command files for those scripts which need to use Python2.

Old repository SHA1: dee179c92af5eb5a84e16722a7d5fb421e4503db