Merge pull request #4118 from RosettaCommons/JackMaguire/cxx17functionattributes
Compilers allow for various function attributes to help annotate code functionality. Most attributes are not useful for us in Rosetta, but a few can be useful for development purposes. For example, [[unused]] makes debug-mode compilation easier/cleaner in the early stages of development.

Starting in c++17, compilers are told to ignore attributes that they do not recognize, which greatly increases portability. This PR adds a macro that allows you to write attributes that will only be applied if your version of c++ is c++17 or later, along with 4 whitelisted attributes: PURE, CONST, UNUSED, and NODISCARD

Old repository SHA1: a344719efdaf62331a97b569a153371d8ee2f299