Merge pull request #4170 from RosettaCommons/roccomoretti/fix_broker_test
Fix issue with silent file endianess flipping on Mac.

The broker integration test was broken related to the fact that the input silent file was created on a big-endian system. Normally that should be fine, as the binary silent file reading code has checks to flip the endianness of the input when read on an alternate-endian system. However, it looks like those checks weren't working properly on the Mac/Clang test server system. 

Best I can tell, the Mac/Clang compiler isn't realizing that the bit twiddling we do to flip endianness will change the memory layout, and is effectively optimizing out the check. I'm not exactly sure why that is, but moving the code body from an inlined function to a cc function seems to fix the issue.

Old repository SHA1: 010c62e53d1698138867069cdacd022403fb69e4